third_party/WebKit/LayoutTests/http/tests/devtools/elements/shadow/update-shadowdom-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
This test confirms that updating the shadow dom is reflected to the Inspector.

- <div id="container">
    - <div id="host">
          #shadow-root (open)
      </div>
  </div>
